Convert ha-entity-toggle to Lit/TS (#2639)
* Convert ha-entity-toggle to Lit/TS

* Update src/components/entity/ha-entity-toggle.ts
